refactor: divide style from component
